rename base->abstract and expose in import
