+++ pymc4/distributions/abstract/__init__.py
