+++ pymc4/distributions/abstract/continuous.py
