+++ pymc4/distributions/abstract/discrete.py
