+++ pymc4/distributions/abstract/distribution.py
